+++ packages/modeler/src/index.ts
@@ -12,5 +12,6 @@ export {
   computeAutoLayout,
   type AutoLayoutOptions,
 } from "./auto-layout.js";
+export { mergeModels, type MergeOptions } from "./merge.js";
 export { ScanModeler, SphereModeler } from "./modeler.js";
 export { default } from "./modeler.js";